server/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.9.0
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
psycopg2-binary==2.9.9

